requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.6
sentence-transformers==3.0.1
transformers==4.42.4
torch==2.0.1
neo4j-driver==5.9.0
llama_cpp_python==0.2.90
pandas==2.2.3
numpy==1.26.4
scipy==1.10.1
jieba==0.42.1
regex==2023.12.25